==> ResNextATAC/5_predict_cell_atlas/human_cell_landscape/4_cluster_refine.ipynb <==
import h5py
import pysam

import numpy as np
import pandas as pd
from scipy.sparse import coo_matrix,csc_matrix

import seaborn as sns
import matplotlib.pyplot as plt
import scanpy as sc
import os
import anndata

mat = csc_matrix(arg1=(50040,100))
ad = sc.AnnData(mat)

ad

anno=pd.read_csv('./meta_pred.csv')
anno

ad.obs_names=anno.barcode.values

ad.obs_names

tsne=pd.read_csv('./tsne_pred.csv',index_col=0)
tsne

tsne1=tsne.loc[anno.barcode.values,]

ad.obsm['X_tsne'] = np.array(tsne1)

ad.obs['lineage'] = anno.lineage.values

ad.obs['subanno'] = anno.subanno.values

ad.obs

cluster=pd.read_csv('./meta_cluster_pred.csv')
cluster

ad.obs['cluster'] = cluster.seurat_clusters.values.astype(str)

sc.settings.verbosity = 3
sc.settings.set_figure_params(dpi=300, facecolor='white')

col=pd.read_csv('/file/path/prefix/color.csv')

sc.pl.tsne(ad, color=['cluster'],legend_loc='on data',legend_fontsize=4,palette=list(col['col'].values),save='_pred')

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#DC143C','#808080','#808080','#808080','#808080','#808080','#808080','#808080','#808080','#808080'])

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#DC143C','#808080','#808080','#808080','#808080','#808080','#808080','#808080','#808080'],save='_epi')

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#808080','#DC143C','#808080','#808080','#808080','#808080','#808080','#808080','#808080'])

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#808080','#808080','#DC143C','#808080','#808080','#808080','#808080','#808080','#808080'],save='_hepa')

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#808080','#808080','#808080','#DC143C','#808080','#808080','#808080','#808080','#808080'],save='_immune')

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#808080','#808080','#808080','#808080','#DC143C','#808080','#808080','#808080','#808080'])

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#808080','#808080','#808080','#808080','#808080','#DC143C','#808080','#808080','#808080'],save='_neuron')

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#808080','#808080','#808080','#808080','#808080','#808080','#DC143C','#808080','#808080'])

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#808080','#808080','#808080','#808080','#808080','#808080','#808080','#DC143C','#808080'])

sc.pl.tsne(ad, color=['lineage'],legend_fontsize=4,palette=['#808080','#808080','#808080','#808080','#808080','#808080','#808080','#808080','#808080','#DC143C'])

sc.pl.tsne(ad, color=['lineage'],palette=list(col['col'].values))
